requirements.txt
aioredis
asyncio
python-dotenv
requests